clean meteor file structure
Created a file structure provided by the MeteorJS Documentation. Added pages that we plan to implement with the following file types:
    - page.html
    - page.css
    - page.sass
    - page.js

The .gitignore had to be edited to remove the /server, for the file structure purposes.

The next goal is to install all of the required packages and begin coding.